requirements.txt
pyautogui
pydirectinput
pycryptodome
psutil